Refactor error and loading selectors
- Move the error and loading selectors into the error and loading
reducers respectively as this is considered a Redux best practice.

- Fix components to import the error and loading selectors correctly.

- Fix the circular progress component in the ThemeButton component to be
in the center.
