+++ client/src/reducers/loading.js
@@ -1,3 +1,15 @@
+/**
+ * A higher order function that returns a single loading status for a single or
+ * multiple API requests (actions). The function returns true if some of the
+ * actions are still pending, and false otherwise.
+ * @param {string[]} actions - list of API requests to check
+ * @param {Object} state - the redux state
+ * @returns {boolean} - true if some API request not complete, false otherwise
+ */
+export const createLoadingSelector = actions => state => {
+  return actions.some(action => state.loading[action]);
+};
+
 export default function loadingReducer(state = {}, action) {
   const { type } = action;
   const matches = /(.*)_(REQUEST|SUCCESS|FAILURE)/.exec(type);
